+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -24,6 +24,7 @@
 - Changed `BodyStream` parameter of `Append` function from pointer to reference.
 - Removed `PathRenameMode`, `PathGetPropertiesAction`, `PathSetAccessControlRecursiveMode`, `FileSystemResourceType` and `FileSystemResourceType`.
 - Removed `IsAccessTierInferred` and `AccessTierChangedOn` from `PathProperties`.
+- Renamed `LeaseDurationType` to `LeaseDuration`, `LeaseStateType` to `LeaseState` and `LeaseStatusType` to `LeaseStatus`.
 
 ## 12.0.0-beta.8 (2021-02-12)