+++ sdk/storage/azure-storage-files-shares/CHANGELOG.md
@@ -29,7 +29,11 @@
 - Renamed `ShareRetentionPolicy` to `RetentionPolicy`.
 - Renamed `ShareProtocolSettings` to `ProtocolSettings`.
 - Renamed `CopyStatusType` to `CopyStatus`
-- Removed `FileRangeWriteType`, `ShareFileRangeList` and `ShareStats`.
+- Removed `FileRangeWriteType`, `ShareFileRangeList`, `FileRangeWriteFromUrlType` and `ShareStats`.
+- Renamed `LeaseDurationType` to `LeaseDuration`, `LeaseStateType` to `LeaseState` and `LeaseStatusType` to `LeaseStatus`.
+- Renamed `ListSharesIncludeType` to `ListSharesIncludeFlags`.
+- Renamed `DeleteSnapshotsOptionType` to `DeleteSnapshotsOption`.
+- Renamed `PermissionCopyModeType` to `PermissionCopyMode`.
 
 ## 12.0.0-beta.8 (2021-02-12)